pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "austack"
version = "0.1.0"
description = "Audio conversation stack with speech-to-text, LLM, and text-to-speech"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "deepgram-sdk>=3.0.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.8.0",
    "baml-py==0.204.0",
    "pyaudio>=0.2.11",
    "webrtcvad>=2.0.10",
    "numpy>=1.24.0",
    "websocket-client>=1.6.0",
    "setuptools>=75.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
austack-server = "austack.server.app:main"

[project.urls]
Homepage = "https://github.com/beam-cloud/austack"
Repository = "https://github.com/beam-cloud/austack.git"
Issues = "https://github.com/beam-cloud/austack/issues"

[tool.hatch.build.targets.wheel]
packages = ["austack"]

[tool.ruff]
line-length = 140
target-version = "py38"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "ISC", "PIE", "PT", "SIM"]
ignore = ["E501", "I001"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
